pid-django-server/pip_requirements.txt
django
django-annoying
numpy==1.15.0
scipy==1.1.0
pandas==0.23.3
matplotlib==2.2.2
six==1.11.0
